新增订单管理

+++ eladmin-system/src/main/java/me/zhengjie/modules/pay/entity/PayMerchant.java
@@ -15,7 +15,7 @@ public class PayMerchant extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long merchantId;
+    private int merchantId;
 
     @Column(name = "md5_secret_key")
     private String md5SecretKey;
